WithIdCursorWrapper For Android SQLite
Subclass of Android CursorWrapper, which adds an _id column to the
result set of a cursor if there is not one already. If not already
present in the result set, the _id will default to the first column, or,
using the optional constructor, can be specified as the last column.
The value of the column will be the position of the row in the result
set.
_id column is required by the CursorAdaptor for a ListFragment; this
class is a means of providing it if not already present.